fp-updater/requirements.txt
aiohttp
aiopath
pillow